refactor-map * Remove "loading"-functionality from DashboardWrapper and a lot of code not needed anymore
