+++ src/dashboards/BusStop/BusStopDashboard.tsx
@@ -1,7 +1,6 @@
 import React from 'react'
 import { DashboardWrapper } from '../../containers/DashboardWrapper/DashboardWrapper'
 import { useAllStopPlaceIds } from '../../logic/use-all-stop-place-ids/useAllStopPlaceIds'
-import { StopPlaceWithDepartures } from '../../types'
 import { BusStopTile } from './components/BusStopTile/BusStopTile'
 import './BusStopDashboard.scss'
 
@@ -9,11 +8,7 @@ const BusStopDashboard = (): JSX.Element | null => {
     const { allStopPlaceIds } = useAllStopPlaceIds()
 
     return (
-        <DashboardWrapper
-            className="bus-stop"
-            // Workaround to bypass loading/no stop found
-            stopPlacesWithDepartures={[{} as StopPlaceWithDepartures]}
-        >
+        <DashboardWrapper className="bus-stop">
             <div>
                 {allStopPlaceIds.map((stopPlaceId) => (
                     <div key={stopPlaceId}>
