+++ src/dashboards/Chrono/ChronoDashboard.tsx
@@ -1,7 +1,6 @@
 import React, { useState, useMemo } from 'react'
 import { WidthProvider, Responsive, Layouts, Layout } from 'react-grid-layout'
 import { useLocation } from 'react-router-dom'
-import { Loader } from '@entur/loader'
 import { useRentalStations, useMobility } from '../../logic'
 import { DashboardWrapper } from '../../containers/DashboardWrapper/DashboardWrapper'
 import { BREAKPOINTS } from '../../constants'
@@ -16,11 +15,11 @@ import { isMobileWeb } from '../../utils/utils'
 import { WeatherTile } from '../../components/WeatherTile/WeatherTile'
 import { ImageTile } from '../../components/ImageTile/ImageTile'
 import { BikeTile } from '../../components/BikeTile/BikeTile'
+import { MapTile } from '../../components/MapTile/MapTile'
 import { FormFactor } from '../../../graphql-generated/mobility-v2'
 import { MobileAppQRTile } from '../../components/QRTile/MobileAppQRTile'
 import { useAllStopPlaceIds } from '../../logic/use-all-stop-place-ids/useAllStopPlaceIds'
 import { ChronoDepartureTile } from './ChronoDepartureTile/ChronoDepartureTile'
-import { MapTile } from './MapTile/MapTile'
 import './ChronoDashboard.scss'
 
 const ResponsiveReactGridLayout = WidthProvider(Responsive)
@@ -97,22 +96,6 @@ const ChronoDashboard = (): JSX.Element | null => {
     const mapCol = settings.showMap ? 1 : 0
     const weatherCol = settings.showWeather ? 1 : 0
 
-    const stopPlacesHasLoaded = Boolean(
-        allStopPlaceIds || settings.hiddenModes.includes('kollektiv'),
-    )
-
-    const bikeHasLoaded = Boolean(
-        bikeRentalStations || settings.hiddenModes.includes('bysykkel'),
-    )
-
-    const scooterHasLoaded = Boolean(
-        scooters || settings.hiddenModes.includes('sparkesykkel'),
-    )
-
-    const hasFetchedData = Boolean(
-        stopPlacesHasLoaded && bikeHasLoaded && scooterHasLoaded,
-    )
-
     const imageTilesToDisplay = useMemo(
         () =>
             settings.showCustomTiles
@@ -142,91 +125,124 @@ const ChronoDashboard = (): JSX.Element | null => {
     )
 
     return (
-        <DashboardWrapper
-            className="chrono"
-            bikeRentalStations={bikeRentalStations}
-        >
-            {!hasFetchedData ? (
-                <div className="compact__loading-screen">
-                    <Loader>Laster inn</Loader>
-                </div>
-            ) : (
-                <div className="chrono__tiles">
-                    <ResponsiveReactGridLayout
-                        key={breakpoint}
-                        breakpoints={BREAKPOINTS}
-                        cols={COLS}
-                        layouts={gridLayouts}
-                        isResizable={!isMobile}
-                        isDraggable={!isMobile}
-                        onBreakpointChange={(newBreakpoint: string) => {
-                            setBreakpoint(newBreakpoint)
-                        }}
-                        onLayoutChange={(
-                            layout: Layout[],
-                            layouts: Layouts,
-                        ): void => {
-                            if (numberOfStopPlaces > 0) {
-                                setGridLayouts(layouts)
-                                saveToLocalStorage(
-                                    dashboardKey as string,
-                                    layouts,
-                                )
-                            }
-                        }}
-                    >
-                        {settings.showMobileAppQrTile && (
-                            <div
-                                key="qr"
-                                data-grid={getDataGrid(
-                                    maxWidthCols - 1,
-                                    maxWidthCols,
-                                    1.8,
-                                    1.8,
-                                    Infinity,
-                                )}
-                            >
-                                <MobileAppQRTile />
-                            </div>
-                        )}
+        <DashboardWrapper className="chrono">
+            <div className="chrono__tiles">
+                <ResponsiveReactGridLayout
+                    key={breakpoint}
+                    breakpoints={BREAKPOINTS}
+                    cols={COLS}
+                    layouts={gridLayouts}
+                    isResizable={!isMobile}
+                    isDraggable={!isMobile}
+                    onBreakpointChange={(newBreakpoint: string) => {
+                        setBreakpoint(newBreakpoint)
+                    }}
+                    onLayoutChange={(
+                        layout: Layout[],
+                        layouts: Layouts,
+                    ): void => {
+                        if (numberOfStopPlaces > 0) {
+                            setGridLayouts(layouts)
+                            saveToLocalStorage(dashboardKey as string, layouts)
+                        }
+                    }}
+                >
+                    {settings.showMobileAppQrTile && (
+                        <div
+                            key="qr"
+                            data-grid={getDataGrid(
+                                maxWidthCols - 1,
+                                maxWidthCols,
+                                1.8,
+                                1.8,
+                                Infinity,
+                            )}
+                        >
+                            <MobileAppQRTile />
+                        </div>
+                    )}
 
-                        {settings.showWeather && (
-                            <div
-                                key="weather"
-                                data-grid={getDataGrid(0, maxWidthCols, 2, 1)}
-                            >
+                    {settings.showWeather && (
+                        <div
+                            key="weather"
+                            data-grid={getDataGrid(0, maxWidthCols, 2, 1)}
+                        >
+                            <ResizeHandle
+                                size="32"
+                                className="resizeHandle"
+                                variant="light"
+                            />
+                            <WeatherTile className="tile" />
+                        </div>
+                    )}
+                    {allStopPlaceIds.map((stopPlaceId, index) => (
+                        <div
+                            key={stopPlaceId}
+                            data-grid={getDataGrid(
+                                weatherCol + index,
+                                maxWidthCols,
+                            )}
+                        >
+                            <ResizeHandle
+                                size="32"
+                                className="resizeHandle"
+                                variant="light"
+                            />
+                            <ChronoDepartureTile stopPlaceId={stopPlaceId} />
+                        </div>
+                    ))}
+                    {bikeRentalStations && anyBikeRentalStations ? (
+                        <div
+                            key="city-bike"
+                            data-grid={getDataGrid(
+                                numberOfStopPlaces + weatherCol,
+                                maxWidthCols,
+                            )}
+                        >
+                            {!isMobile ? (
                                 <ResizeHandle
                                     size="32"
                                     className="resizeHandle"
                                     variant="light"
                                 />
-                                <WeatherTile className="tile" />
-                            </div>
-                        )}
-                        {allStopPlaceIds.map((stopPlaceId, index) => (
-                            <div
-                                key={stopPlaceId}
-                                data-grid={getDataGrid(
-                                    weatherCol + index,
-                                    maxWidthCols,
-                                )}
-                            >
+                            ) : null}
+                            <BikeTile stations={bikeRentalStations} />
+                        </div>
+                    ) : (
+                        []
+                    )}
+                    {hasData && mapCol && (
+                        <div
+                            id="chrono-map-tile"
+                            key="map"
+                            data-grid={getDataGrid(
+                                numberOfStopPlaces + bikeCol + weatherCol,
+                                maxWidthCols,
+                            )}
+                        >
+                            {!isMobile && (
                                 <ResizeHandle
                                     size="32"
                                     className="resizeHandle"
-                                    variant="light"
+                                    variant="dark"
                                 />
-                                <ChronoDepartureTile
-                                    stopPlaceId={stopPlaceId}
-                                />
-                            </div>
-                        ))}
-                        {bikeRentalStations && anyBikeRentalStations ? (
+                            )}
+                            <MapTile />
+                        </div>
+                    )}
+                    {imageTilesToDisplay.length > 0 &&
+                        imageTilesToDisplay.map((imageTile, index) => (
                             <div
-                                key="city-bike"
+                                key={imageTile.id}
                                 data-grid={getDataGrid(
-                                    numberOfStopPlaces + weatherCol,
+                                    numberOfStopPlaces +
+                                        weatherCol +
+                                        bikeCol +
+                                        mapCol +
+                                        index,
                                     maxWidthCols,
+                                    10,
+                                    2,
                                 )}
                             >
                                 {!isMobile ? (
@@ -236,53 +252,26 @@ const ChronoDashboard = (): JSX.Element | null => {
                                         variant="light"
                                     />
                                 ) : null}
-                                <BikeTile stations={bikeRentalStations} />
+                                <ImageTile {...imageTile} />
                             </div>
-                        ) : (
-                            []
-                        )}
-                        {hasData && mapCol ? (
+                        ))}
+                    {qrTilesToDisplay.length > 0 &&
+                        qrTilesToDisplay.map((qrTile, index) => (
                             <div
-                                id="chrono-map-tile"
-                                key="map"
+                                key={qrTile.id}
                                 data-grid={getDataGrid(
-                                    numberOfStopPlaces + bikeCol + weatherCol,
+                                    numberOfStopPlaces +
+                                        weatherCol +
+                                        bikeCol +
+                                        mapCol +
+                                        numberOfCustomImages +
+                                        index,
                                     maxWidthCols,
+                                    10,
+                                    3,
                                 )}
                             >
-                                {!isMobile ? (
-                                    <ResizeHandle
-                                        size="32"
-                                        className="resizeHandle"
-                                        variant="dark"
-                                    />
-                                ) : null}
-                                <MapTile
-                                    scooters={scooters}
-                                    bikeRentalStations={bikeRentalStations}
-                                    latitude={settings.coordinates.latitude}
-                                    longitude={settings.coordinates.longitude}
-                                    zoom={settings.zoom}
-                                />
-                            </div>
-                        ) : (
-                            []
-                        )}
-                        {imageTilesToDisplay.length > 0 &&
-                            imageTilesToDisplay.map((imageTile, index) => (
-                                <div
-                                    key={imageTile.id}
-                                    data-grid={getDataGrid(
-                                        numberOfStopPlaces +
-                                            weatherCol +
-                                            bikeCol +
-                                            mapCol +
-                                            index,
-                                        maxWidthCols,
-                                        10,
-                                        2,
-                                    )}
-                                >
+                                <div className="tile">
                                     {!isMobile ? (
                                         <ResizeHandle
                                             size="32"
@@ -290,40 +279,12 @@ const ChronoDashboard = (): JSX.Element | null => {
                                             variant="light"
                                         />
                                     ) : null}
-                                    <ImageTile {...imageTile} />
+                                    <QRBox {...qrTile} />
                                 </div>
-                            ))}
-                        {qrTilesToDisplay.length > 0 &&
-                            qrTilesToDisplay.map((qrTile, index) => (
-                                <div
-                                    key={qrTile.id}
-                                    data-grid={getDataGrid(
-                                        numberOfStopPlaces +
-                                            weatherCol +
-                                            bikeCol +
-                                            mapCol +
-                                            numberOfCustomImages +
-                                            index,
-                                        maxWidthCols,
-                                        10,
-                                        3,
-                                    )}
-                                >
-                                    <div className="tile">
-                                        {!isMobile ? (
-                                            <ResizeHandle
-                                                size="32"
-                                                className="resizeHandle"
-                                                variant="light"
-                                            />
-                                        ) : null}
-                                        <QRBox {...qrTile} />
-                                    </div>
-                                </div>
-                            ))}
-                    </ResponsiveReactGridLayout>
-                </div>
-            )}
+                            </div>
+                        ))}
+                </ResponsiveReactGridLayout>
+            </div>
         </DashboardWrapper>
     )
 }
