--- src/dashboards/Chrono/MapTile/MapTile.scss
@@ -1,34 +1,0 @@
-.chrono .maptile {
-    height: inherit;
-    width: inherit;
-    background-color: var(--tavla-box-background-color);
-    color: var(--tavla-font-color);
-    flex: none;
-    overflow: auto;
-    overflow-x: hidden;
-    padding-top: 0;
-
-    &__header {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        margin-bottom: 1rem;
-        padding: 0 2rem;
-
-        h2 {
-            margin: 0;
-            color: var(--tavla-font-color);
-            font-size: 2rem;
-        }
-    }
-
-    &__header-icons {
-        display: flex;
-        align-items: center;
-        color: var(--colors-blues-blue60);
-    }
-
-    @media (max-width: 1000px) {
-        height: 39.375rem;
-    }
-}
