+++ src/dashboards/Compact/CompactDashboard.tsx
@@ -1,8 +1,7 @@
 import React, { useState, useMemo } from 'react'
 import { WidthProvider, Responsive, Layouts, Layout } from 'react-grid-layout'
 import { useLocation } from 'react-router-dom'
-import { Loader } from '@entur/loader'
-import { useRentalStations, useMobility } from '../../logic'
+import { useRentalStations } from '../../logic'
 import { useAllStopPlaceIds } from '../../logic/use-all-stop-place-ids/useAllStopPlaceIds'
 import { DashboardWrapper } from '../../containers/DashboardWrapper/DashboardWrapper'
 import { ResizeHandle } from '../../assets/icons/ResizeHandle'
@@ -16,10 +15,10 @@ import { isMobileWeb } from '../../utils/utils'
 import { WeatherTile } from '../../components/WeatherTile/WeatherTile'
 import { ImageTile } from '../../components/ImageTile/ImageTile'
 import { BikeTile } from '../../components/BikeTile/BikeTile'
+import { MapTile } from '../../components/MapTile/MapTile'
 import { FormFactor } from '../../../graphql-generated/mobility-v2'
 import { MobileAppQRTile } from '../../components/QRTile/MobileAppQRTile'
 import { CompactDepartureTile } from './CompactDepartureTile/CompactDepartureTile'
-import { MapTile } from './MapTile/MapTile'
 import './CompactDashboard.scss'
 
 const ResponsiveReactGridLayout = WidthProvider(Responsive)
@@ -77,7 +76,6 @@ const CompactDashboard = (): JSX.Element | null => {
         FormFactor.Bicycle,
         settings.hiddenModes.includes('bysykkel'),
     )
-    const scooters = useMobility(FormFactor.Scooter)
 
     const { allStopPlaceIds } = useAllStopPlaceIds()
 
@@ -89,19 +87,7 @@ const CompactDashboard = (): JSX.Element | null => {
     const mapCol = settings.showMap ? 1 : 0
     const weatherCol = settings.showWeather ? 1 : 0
 
-    const hasData = Boolean(bikeRentalStations?.length || scooters?.length)
-
     const maxWidthCols = COLS[breakpoint] || 1
-
-    const bikeHasLoaded = Boolean(
-        bikeRentalStations || settings.hiddenModes.includes('bysykkel'),
-    )
-
-    const scooterHasLoaded = Boolean(
-        scooters || settings.hiddenModes.includes('sparkesykkel'),
-    )
-
-    const hasFetchedData = Boolean(bikeHasLoaded && scooterHasLoaded)
 
     const imageTilesToDisplay = useMemo(
         () =>
@@ -132,92 +118,122 @@ const CompactDashboard = (): JSX.Element | null => {
     )
 
     return (
-        <DashboardWrapper
-            className="compact"
-            bikeRentalStations={bikeRentalStations}
-            scooters={scooters}
-        >
-            {!hasFetchedData ? (
-                <div className="compact__loading-screen">
-                    <Loader>Laster inn</Loader>
-                </div>
-            ) : (
-                <div className="compact__tiles">
-                    <ResponsiveReactGridLayout
-                        key={breakpoint}
-                        breakpoints={BREAKPOINTS}
-                        cols={COLS}
-                        layouts={gridLayouts}
-                        isResizable={!isMobile}
-                        isDraggable={!isMobile}
-                        margin={[32, 32]}
-                        onBreakpointChange={(newBreakpoint: string) => {
-                            setBreakpoint(newBreakpoint)
-                        }}
-                        onLayoutChange={(
-                            layout: Layout[],
-                            layouts: Layouts,
-                        ): void => {
-                            if (numberOfStopPlaces > 0) {
-                                setGridLayouts(layouts)
-                                saveToLocalStorage(
-                                    dashboardKey as string,
-                                    layouts,
-                                )
-                            }
-                        }}
-                    >
-                        {settings.showMobileAppQrTile && (
-                            <div
-                                key="qr"
-                                data-grid={getDataGrid(
-                                    maxWidthCols - 1,
-                                    maxWidthCols,
-                                    1.8,
-                                    1.8,
-                                    Infinity,
-                                )}
-                            >
-                                <MobileAppQRTile />
-                            </div>
-                        )}
-                        {settings.showWeather && (
-                            <div
-                                key="weather"
-                                data-grid={getDataGrid(0, maxWidthCols, 1, 1)}
-                            >
+        <DashboardWrapper className="compact">
+            <div className="compact__tiles">
+                <ResponsiveReactGridLayout
+                    key={breakpoint}
+                    breakpoints={BREAKPOINTS}
+                    cols={COLS}
+                    layouts={gridLayouts}
+                    isResizable={!isMobile}
+                    isDraggable={!isMobile}
+                    margin={[32, 32]}
+                    onBreakpointChange={(newBreakpoint: string) => {
+                        setBreakpoint(newBreakpoint)
+                    }}
+                    onLayoutChange={(
+                        layout: Layout[],
+                        layouts: Layouts,
+                    ): void => {
+                        if (numberOfStopPlaces > 0) {
+                            setGridLayouts(layouts)
+                            saveToLocalStorage(dashboardKey as string, layouts)
+                        }
+                    }}
+                >
+                    {settings.showMobileAppQrTile && (
+                        <div
+                            key="qr"
+                            data-grid={getDataGrid(
+                                maxWidthCols - 1,
+                                maxWidthCols,
+                                1.8,
+                                1.8,
+                                Infinity,
+                            )}
+                        >
+                            <MobileAppQRTile />
+                        </div>
+                    )}
+                    {settings.showWeather && (
+                        <div
+                            key="weather"
+                            data-grid={getDataGrid(0, maxWidthCols, 1, 1)}
+                        >
+                            <ResizeHandle
+                                size="32"
+                                className="resizeHandle"
+                                variant="light"
+                            />
+                            <WeatherTile className="tile" />
+                        </div>
+                    )}
+                    {allStopPlaceIds.map((stopPlaceId, index) => (
+                        <div
+                            key={stopPlaceId}
+                            data-grid={getDataGrid(
+                                weatherCol + index,
+                                maxWidthCols,
+                            )}
+                        >
+                            <ResizeHandle
+                                size="32"
+                                className="resizeHandle"
+                                variant="light"
+                            />
+                            <CompactDepartureTile stopPlaceId={stopPlaceId} />
+                        </div>
+                    ))}
+                    {bikeRentalStations && anyBikeRentalStations && (
+                        <div
+                            key="city-bike"
+                            data-grid={getDataGrid(
+                                numberOfStopPlaces + weatherCol,
+                                maxWidthCols,
+                            )}
+                        >
+                            {!isMobile ? (
                                 <ResizeHandle
                                     size="32"
                                     className="resizeHandle"
                                     variant="light"
                                 />
-                                <WeatherTile className="tile" />
-                            </div>
-                        )}
-                        {allStopPlaceIds.map((stopPlaceId, index) => (
-                            <div
-                                key={stopPlaceId}
-                                data-grid={getDataGrid(
-                                    weatherCol + index,
-                                    maxWidthCols,
-                                )}
-                            >
+                            ) : null}
+                            <BikeTile stations={bikeRentalStations} />
+                        </div>
+                    )}
+                    {settings.showMap && (
+                        <div
+                            id="compact-map-tile"
+                            key="map"
+                            data-grid={getDataGrid(
+                                numberOfStopPlaces + bikeCol + weatherCol,
+                                maxWidthCols,
+                            )}
+                        >
+                            {!isMobile && (
                                 <ResizeHandle
                                     size="32"
                                     className="resizeHandle"
-                                    variant="light"
+                                    variant="dark"
                                 />
-                                <CompactDepartureTile
-                                    stopPlaceId={stopPlaceId}
-                                />
-                            </div>
-                        ))}
-                        {bikeRentalStations && anyBikeRentalStations ? (
+                            )}
+                            <MapTile />
+                        </div>
+                    )}
+                    {imageTilesToDisplay.length > 0 &&
+                        imageTilesToDisplay.map((imageTile, index) => (
                             <div
-                                key="city-bike"
+                                key={imageTile.id}
                                 data-grid={getDataGrid(
-                                    numberOfStopPlaces + weatherCol,
+                                    numberOfStopPlaces +
+                                        weatherCol +
+                                        bikeCol +
+                                        mapCol +
+                                        index,
                                     maxWidthCols,
+                                    10,
+                                    2,
                                 )}
                             >
                                 {!isMobile ? (
@@ -227,93 +243,37 @@ const CompactDashboard = (): JSX.Element | null => {
                                         variant="light"
                                     />
                                 ) : null}
-                                <BikeTile stations={bikeRentalStations} />
+                                <ImageTile {...imageTile} />
                             </div>
-                        ) : (
-                            []
-                        )}
-                        {hasData && mapCol ? (
+                        ))}
+                    {qrTilesToDisplay.length > 0 &&
+                        qrTilesToDisplay.map((qrTile, index) => (
                             <div
-                                id="compact-map-tile"
-                                key="map"
+                                key={qrTile.id}
                                 data-grid={getDataGrid(
-                                    numberOfStopPlaces + bikeCol + weatherCol,
+                                    numberOfStopPlaces +
+                                        weatherCol +
+                                        bikeCol +
+                                        mapCol +
+                                        imageTilesToDisplay.length +
+                                        index,
                                     maxWidthCols,
+                                    10,
+                                    3,
                                 )}
                             >
                                 {!isMobile ? (
                                     <ResizeHandle
                                         size="32"
                                         className="resizeHandle"
-                                        variant="dark"
+                                        variant="light"
                                     />
                                 ) : null}
-
-                                <MapTile
-                                    scooters={scooters}
-                                    bikeRentalStations={bikeRentalStations}
-                                    latitude={settings.coordinates.latitude}
-                                    longitude={settings.coordinates.longitude}
-                                    zoom={settings.zoom}
-                                />
+                                <div className="tile"></div>
                             </div>
-                        ) : (
-                            []
-                        )}
-                        {imageTilesToDisplay.length > 0 &&
-                            imageTilesToDisplay.map((imageTile, index) => (
-                                <div
-                                    key={imageTile.id}
-                                    data-grid={getDataGrid(
-                                        numberOfStopPlaces +
-                                            weatherCol +
-                                            bikeCol +
-                                            mapCol +
-                                            index,
-                                        maxWidthCols,
-                                        10,
-                                        2,
-                                    )}
-                                >
-                                    {!isMobile ? (
-                                        <ResizeHandle
-                                            size="32"
-                                            className="resizeHandle"
-                                            variant="light"
-                                        />
-                                    ) : null}
-                                    <ImageTile {...imageTile} />
-                                </div>
-                            ))}
-                        {qrTilesToDisplay.length > 0 &&
-                            qrTilesToDisplay.map((qrTile, index) => (
-                                <div
-                                    key={qrTile.id}
-                                    data-grid={getDataGrid(
-                                        numberOfStopPlaces +
-                                            weatherCol +
-                                            bikeCol +
-                                            mapCol +
-                                            imageTilesToDisplay.length +
-                                            index,
-                                        maxWidthCols,
-                                        10,
-                                        3,
-                                    )}
-                                >
-                                    {!isMobile ? (
-                                        <ResizeHandle
-                                            size="32"
-                                            className="resizeHandle"
-                                            variant="light"
-                                        />
-                                    ) : null}
-                                    <div className="tile"></div>
-                                </div>
-                            ))}
-                    </ResponsiveReactGridLayout>
-                </div>
-            )}
+                        ))}
+                </ResponsiveReactGridLayout>
+            </div>
         </DashboardWrapper>
     )
 }
