--- src/dashboards/Compact/MapTile/MapTile.scss
@@ -1,30 +1,0 @@
-.compact .maptile {
-    height: inherit;
-    width: inherit;
-    background-color: var(--tavla-box-background-color);
-    flex: none;
-    overflow: scroll;
-    overflow-x: hidden;
-    padding-top: 0;
-
-    &__header {
-        display: flex;
-        justify-content: space-between;
-        padding: 0 2rem;
-        align-items: center;
-        margin-bottom: 1rem;
-
-        h2 {
-            margin: 0;
-            color: var(--tavla-font-color);
-            font-size: 2rem;
-        }
-    }
-
-    &__header-icons {
-        display: flex;
-        align-items: center;
-        font-size: 2rem;
-        color: var(--colors-blues-blue60);
-    }
-}
