+++ src/dashboards/Map/DepartureIcon/DepartureIcon.module.scss
@@ -10,7 +10,6 @@
     display: flex;
     font-size: Variables.$font-sizes-medium;
     padding: 0.3rem 0.5625rem 0 0.5625rem;
-
 }
 
 .Icon {
