+++ src/dashboards/Map/DepartureTag/DepartureTag.module.scss
@@ -15,7 +15,7 @@
 .Stop {
     margin: 1rem 1.5rem 0.5rem 1.5rem;
     font-size: 1.25rem;
-    color: Variables.$colors-brand-blue
+    color: Variables.$colors-brand-blue;
 }
 
 .DepartureRow {
@@ -25,7 +25,6 @@
     margin-left: 1.5rem;
     margin-bottom: 1.5rem;
     padding-top: 0.5rem;
-
 }
 
 .Date {
