+++ src/dashboards/Timeline/TimelineDashboard.tsx
@@ -1,5 +1,4 @@
 import React from 'react'
-import { StopPlaceWithDepartures } from '../../types'
 import { DashboardWrapper } from '../../containers/DashboardWrapper/DashboardWrapper'
 import { useAllStopPlaceIds } from '../../logic/use-all-stop-place-ids/useAllStopPlaceIds'
 import { TimelineTile } from './TimelineTile/TimelineTile'
@@ -9,10 +8,7 @@ const TimelineDashboard = (): JSX.Element | null => {
     const { allStopPlaceIds } = useAllStopPlaceIds()
 
     return (
-        <DashboardWrapper
-            className={classes.TimelineDashboard}
-            stopPlacesWithDepartures={[{} as StopPlaceWithDepartures]}
-        >
+        <DashboardWrapper className={classes.TimelineDashboard}>
             {allStopPlaceIds.map((stopPlaceId) => (
                 <TimelineTile key={stopPlaceId} stopPlaceId={stopPlaceId} />
             ))}
